requirements.txt
setuptools
python-dotenv
numpy
pandas
pymongo
certifi
pymongo[srv]==4.13
scikit-learn


# -e .